% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-posterior-predict.R
\name{.stan_posterior_predict}
\alias{.stan_posterior_predict}
\title{Underlying function to run posterior predictions with terra.}
\usage{
.stan_posterior_predict(mod, dat, ...)
}
\arguments{
\item{mod}{An rstanarm fitted model object.}

\item{dat}{Input data to fit the model, which then gets passed as 'newdata'
into \code{rstanarm::posterior_predict}.}

\item{...}{Extra arguments passed into \code{rstanarm::posterior_predict} (such as
'draws').}
}
\value{
A matrix of replicated values with dimensions \code{nrow(dat)} by \code{draws}
(note this is the transposed output of \code{rstanarm::posterior_predict}).
}
\description{
Underlying function to run posterior predictions with terra.
}
