% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terra-posterior-predict.R
\name{terra_posterior_predict}
\alias{terra_posterior_predict}
\title{Simulate from the posterior predictive distribution with terra.}
\usage{
terra_posterior_predict(object, model, draws = 1, cores = 1)
}
\arguments{
\item{object}{A spatial object (e.g., \code{SpatRaster}) that serves as the
covariate data for prediction. Passed to \code{\link[terra:predict]{terra::predict()}}.}

\item{model}{A fitted Stan model object, typically returned from a cmdstanr
modeling workflow.}

\item{draws}{Integer specifying the number of posterior draws to use in
prediction. Defaults to 1.}

\item{cores}{Integer specifying the number of CPU cores to use. Currently,
only \code{cores = 1} is implemented.}
}
\value{
A \code{SpatRaster} object with posterior predictive values.
}
\description{
Utilizing a SpatRaster to hold the covariates, simulate the posterior
predictive distribution (PPD) of a model and return a SpatRaster whose layers
are draws from the PPD.
}
